Move status logic into troubleshoot command file
The Status is used only in the troubleshoot command.
