+++ cmd/troubleshoot.go
@@ -1,8 +1,13 @@
 package cmd
 
 import (
+	"bytes"
 	"fmt"
+	"html/template"
 	"net/http"
+	"runtime"
+	"strings"
+	"sync"
 	"time"
 
 	"github.com/exercism/cli/cli"
@@ -32,7 +37,7 @@ command into a GitHub issue so we can help figure out what's going on.
 			return err
 		}
 
-		status := cli.NewStatus(c, *cfg)
+		status := NewStatus(c, *cfg)
 		status.Censor = !fullAPIKey
 		s, err := status.Check()
 		if err != nil {
@@ -43,6 +48,206 @@ command into a GitHub issue so we can help figure out what's going on.
 		return nil
 	},
 }
+
+// Status represents the results of a CLI self test.
+type Status struct {
+	Censor          bool
+	Version         versionStatus
+	System          systemStatus
+	Configuration   configurationStatus
+	APIReachability apiReachabilityStatus
+	cli             *cli.CLI
+	cfg             config.UserConfig
+}
+
+type versionStatus struct {
+	Current  string
+	Latest   string
+	Status   string
+	Error    error
+	UpToDate bool
+}
+
+type systemStatus struct {
+	OS           string
+	Architecture string
+	Build        string
+}
+
+type configurationStatus struct {
+	Home      string
+	Workspace string
+	File      string
+	Token     string
+	TokenURL  string
+}
+
+type apiReachabilityStatus struct {
+	Services []*apiPing
+}
+
+type apiPing struct {
+	Service string
+	URL     string
+	Status  string
+	Latency time.Duration
+}
+
+// NewStatus prepares a value to perform a diagnostic self-check.
+func NewStatus(c *cli.CLI, uc config.UserConfig) Status {
+	status := Status{
+		cli: c,
+		cfg: uc,
+	}
+	return status
+}
+
+// Check runs the CLI's diagnostic self-check.
+func (status *Status) Check() (string, error) {
+	status.Version = newVersionStatus(status.cli)
+	status.System = newSystemStatus()
+	status.Configuration = newConfigurationStatus(status)
+	status.APIReachability = newAPIReachabilityStatus(status.cfg.APIBaseURL)
+
+	return status.compile()
+}
+func (status *Status) compile() (string, error) {
+	t, err := template.New("self-test").Parse(tmplSelfTest)
+	if err != nil {
+		return "", err
+	}
+
+	var bb bytes.Buffer
+	t.Execute(&bb, status)
+	return bb.String(), nil
+}
+
+func newAPIReachabilityStatus(baseURL string) apiReachabilityStatus {
+	ar := apiReachabilityStatus{
+		Services: []*apiPing{
+			{Service: "GitHub", URL: "https://api.github.com"},
+			{Service: "Exercism", URL: fmt.Sprintf("%s/ping", baseURL)},
+		},
+	}
+	var wg sync.WaitGroup
+	wg.Add(len(ar.Services))
+	for _, service := range ar.Services {
+		go service.Call(&wg)
+	}
+	wg.Wait()
+	return ar
+}
+
+func newVersionStatus(c *cli.CLI) versionStatus {
+	vs := versionStatus{
+		Current: c.Version,
+	}
+	ok, err := c.IsUpToDate()
+	if err == nil {
+		vs.Latest = c.LatestRelease.Version()
+	} else {
+		vs.Error = fmt.Errorf("Error: %s", err)
+	}
+	vs.UpToDate = ok
+	return vs
+}
+
+func newSystemStatus() systemStatus {
+	ss := systemStatus{
+		OS:           runtime.GOOS,
+		Architecture: runtime.GOARCH,
+	}
+	if cli.BuildOS != "" && cli.BuildARCH != "" {
+		ss.Build = fmt.Sprintf("%s/%s", cli.BuildOS, cli.BuildARCH)
+	}
+	if cli.BuildARM != "" {
+		ss.Build = fmt.Sprintf("%s ARMv%s", ss.Build, cli.BuildARM)
+	}
+	return ss
+}
+
+func newConfigurationStatus(status *Status) configurationStatus {
+	cs := configurationStatus{
+		Home:      status.cfg.Home,
+		Workspace: status.cfg.Workspace,
+		File:      status.cfg.File(),
+		Token:     status.cfg.Token,
+		TokenURL:  config.InferSiteURL(status.cfg.APIBaseURL) + "/my/settings",
+	}
+	if status.Censor && status.cfg.Token != "" {
+		cs.Token = redactToken(status.cfg.Token)
+	}
+	return cs
+}
+
+func (ping *apiPing) Call(wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	now := time.Now()
+	res, err := cli.HTTPClient.Get(ping.URL)
+	delta := time.Since(now)
+	ping.Latency = delta
+	if err != nil {
+		ping.Status = err.Error()
+		return
+	}
+	res.Body.Close()
+	ping.Status = "connected"
+}
+
+func redactToken(token string) string {
+	str := token[4 : len(token)-3]
+	redaction := strings.Repeat("*", len(str))
+	return string(token[:4]) + redaction + string(token[len(token)-3:])
+}
+
+const tmplSelfTest = `
+Troubleshooting Information
+===========================
+
+Version
+----------------
+Current: {{ .Version.Current }}
+Latest:  {{ with .Version.Latest }}{{ . }}{{ else }}<unknown>{{ end }}
+{{ with .Version.Error }}
+{{ . }}
+{{ end -}}
+{{ if not .Version.UpToDate }}
+Call 'exercism upgrade' to get the latest version.
+See the release notes at https://github.com/exercism/cli/releases/tag/{{ .Version.Latest }} for details.
+{{ end }}
+
+Operating System
+----------------
+OS:           {{ .System.OS }}
+Architecture: {{ .System.Architecture }}
+{{ with .System.Build }}
+Build: {{ . }}
+{{ end }}
+
+Configuration
+----------------
+Home:      {{ .Configuration.Home }}
+Workspace: {{ .Configuration.Workspace }}
+Config:    {{ .Configuration.File }}
+API key:   {{ with .Configuration.Token }}{{ . }}{{ else }}<not configured>
+Find your API key at {{ .Configuration.TokenURL }}{{ end }}
+
+API Reachability
+----------------
+{{ range .APIReachability.Services }}
+{{ .Service }}:
+    * {{ .URL }}
+    * [{{ .Status }}]
+    * {{ .Latency }}
+{{ end }}
+
+If you are having trouble please file a GitHub issue at
+https://github.com/exercism/exercism.io/issues and include
+this information.
+{{ if not .Censor }}
+Don't share your API key. Keep that private.
+{{ end }}`
 
 func init() {
 	RootCmd.AddCommand(troubleshootCmd)
