+++ cmd/troubleshoot_test.go
@@ -1,8 +1,9 @@
-package cli
+package cmd
 
 import (
 	"testing"
 
+	"github.com/exercism/cli/cli"
 	"github.com/exercism/cli/config"
 	"github.com/stretchr/testify/assert"
 )
@@ -11,7 +12,7 @@ func TestCensor(t *testing.T) {
 	fakeToken := "1a11111aaaa111aa1a11111a11111aa1"
 	expected := "1a11*************************aa1"
 
-	c := New("")
+	c := cli.New("")
 
 	cfg := config.NewEmptyUserConfig()
 	cfg.Token = fakeToken
